Fix: ST40/ST41 first packet timestamp
Packet timestamps were only calculated when
the RTP time differed from the previous frame,
which failed for the initial packet. This caused
the first TX packet to be transmitted with
timestamp 0, resulting in DDoS-like behavior
on the RX side.

During initalization make the session calculate
the timestamp to avoid such issues.

To test set the date into problematic clock area
sudo date -s "2025-10-29 17:28:30"

+++ lib/src/st2110/st_header.h
@@ -1190,7 +1190,7 @@ struct st_rx_ancillary_session_impl {
 
   struct mt_rtcp_rx* rtcp_rx[MTL_SESSION_PORT_MAX];
 
-  uint32_t tmstamp;
+  int64_t tmstamp;
   /* status */
   rte_atomic32_t stat_frames_received;
   int stat_pkts_dropped;
@@ -1363,7 +1363,8 @@ struct st_rx_fastmetadata_session_impl {
 
   struct mt_rtcp_rx* rtcp_rx[MTL_SESSION_PORT_MAX];
 
-  uint32_t tmstamp;
+  /* the timestamp */
+  int64_t tmstamp;
   /* status */
   rte_atomic32_t stat_frames_received;
   int stat_pkts_redundant;
